// frontend/src/api/generated/model/floor.ts
export interface Floor {
  floor: number;
  /** @minimum 0 */
  rooms_count: number;
  /** @minimum 0 */
  rooms_with_free_beds: number;
  /** @minimum 0 */
  beds_count: number;
  /** @minimum 0 */
  free_beds: number;
}